requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aioschedule==0.5.2
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
MockupEngineer==2022.3.24.1
multidict==6.0.4
Pillow==9.4.0
pycodestyle==2.10.0
pydantic==1.10.5
pytz==2022.7.1
requests==2.28.2
TemporaryStorage==2022.1.31.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
